Добавил поддержку базы данных h2

+++ flowershop/src/main/java/com/accenture/flowershop/be/access/user/UserAccess.java
@@ -3,11 +3,12 @@ package com.accenture.flowershop.be.access.user;
 import com.accenture.flowershop.be.entity.user.User;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Repository
 public interface UserAccess {
     void saveUser(String login, String password, String name, String address,
-                  String phoneNumber, double score, int sale, String role);
+                  String phoneNumber, BigDecimal score, int sale, String role);
     Map<String, User> getAllUsers();
 }
