+++ flowershop/src/main/java/com/accenture/flowershop/be/business/user/UserBusinessService.java
@@ -3,10 +3,12 @@ package com.accenture.flowershop.be.business.user;
 import com.accenture.flowershop.fe.dto.UserDTO;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 public interface UserBusinessService {
     void saveUser(String login, String password, String name, String address,
-                  String phoneNumber, double score, int sale, String role);
+                  String phoneNumber, BigDecimal score, int sale, String role);
 
     UserDTO getUserDTO(String login, String password);
